fix: 아키텍처 위반 - SubmitAnalysisResult/PollAnalysisResult 도메인 타입으로 이동
