+++ src/domain/types.ts
@@ -453,3 +453,16 @@ export type RawAnalysisResponse = Omit<
     strategyResults: Omit<StrategyResult, 'confidenceWeight' | 'id'>[];
     candlePatterns: Omit<CandlePatternSummary, 'id'>[];
 };
+
+// --- Job Result Types (submit + poll 패턴) ---
+
+/** submitAnalysisAction 반환 타입 */
+export type SubmitAnalysisResult =
+    | { status: 'cached'; result: AnalysisResponse; skillsDegraded: boolean }
+    | { status: 'submitted'; jobId: string };
+
+/** pollAnalysisAction 반환 타입 */
+export type PollAnalysisResult =
+    | { status: 'processing' }
+    | { status: 'done'; result: AnalysisResponse; skillsDegraded: boolean }
+    | { status: 'error'; error: string };
