+++ src/infrastructure/jobs/types.ts
@@ -1,4 +1,4 @@
-import type { AnalysisResponse, Timeframe } from '@/domain/types';
+import type { Timeframe } from '@/domain/types';
 
 export type JobStatus = 'processing' | 'done' | 'error';
 
@@ -6,14 +6,3 @@ export interface JobMeta {
     symbol: string;
     timeframe: Timeframe;
 }
-
-/** submitAnalysisAction 반환 타입 */
-export type SubmitAnalysisResult =
-    | { status: 'cached'; result: AnalysisResponse; skillsDegraded: boolean }
-    | { status: 'submitted'; jobId: string };
-
-/** pollAnalysisAction 반환 타입 */
-export type PollAnalysisResult =
-    | { status: 'processing' }
-    | { status: 'done'; result: AnalysisResponse; skillsDegraded: boolean }
-    | { status: 'error'; error: string };
